requirements.txt
nicegui
azure-identity
azure-storage-blob